viz: round rooms now

+++ libft/libft.h
@@ -91,7 +91,7 @@ int				ft_lstsize(t_list *list);
 int				ft_abs(int i);
 long double		ft_abs_float(long double nb);
 int				get_next_line(const int fd, char **line);
-int				ft_sqrt(int nb);
+float			ft_sqrt(int nb);
 char			*ft_itoa_base(int nb, int base);
 char			*ft_ull_itoa_base(unsigned long long int nb, int base);
 char			*ft_ll_itoa(long long int n);
